+++ src/aideas/app/config.py
@@ -630,7 +630,7 @@ class RunArg(str, Enum):
         return self.__path
 
     AGENTS = ('agents', 'a', 'list')
-    BROWSER_VISIBLE = ('browser-visible', 'bv', 'bool', True, False)
+    BROWSER_TYPE = ('browser-mode', 'bm', 'str', True, False)
     CONTINUE_ON_ERROR = ('continue-on-error', 'coe', 'bool', True, False)
     INPUT_LANGUAGE_CODE = ('input-language-code', 'ilc', 'str', False, False)
     IMAGE_FILE_LANDSCAPE = ('image-file-landscape', 'vci', 'str', False, True)