+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.1] - 2025-08-16
+
 ### Added
 
 - Run arg `--save-screens` (`-ss`), a string to specify when screenshots and web pages